Componnets decompose
