+++ src/demo/index.tsx
@@ -7,7 +7,12 @@ import ProseMirror from "./../lib/index";
 
 const root = document.getElementById("react-root");
 
-ReactDOM.render(<ProseMirror />, root);
+ReactDOM.render(
+  <div>
+    <ProseMirror />
+  </div>,
+  root
+);
 
 if (module.hot) {
   module.hot.accept();
