+++ src/lib/components/Editor/cut.ts
