+++ src/lib/components/Editor/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Editor";
