+++ src/lib/components/Popup/Popup.tsx
@@ -7,7 +7,7 @@ import {
   Subscriber,
   ScreenPosition,
   PopupState
-} from "../types";
+} from "../../types";
 
 import "./index.css";
 
@@ -15,22 +15,43 @@ interface Props {
   subscribeToPluginChanges: SubscribHandler;
 }
 
+const defaultProps: PopupState = {
+  isVisible: false,
+  word: "",
+  screenPos: { x: 0, y: 0 },
+  suggestions: [],
+  clickHandler: (correction: string) => {}
+};
+
 const Popup = ({ subscribeToPluginChanges }: Props) => {
   const [
     { isVisible, word, screenPos, suggestions, clickHandler: clickHandlerProp },
     setState
-  ] = useState<PopupState>({
-    isVisible: false,
-    word: "",
-    screenPos: { x: 0, y: 0 },
-    suggestions: [],
-    clickHandler: (correction: string) => {}
-  });
+  ] = useState<PopupState>(defaultProps);
+
+  console.log("popup");
 
   useEffect(() => {
-    const unsubscribe = subscribeToPluginChanges(setState);
+    const unsubscribe = subscribeToPluginChanges(newProps => {
+      const {
+        isVisible: newIsVisible,
+        word: newWord,
+        screenPos: newScerenPos,
+        suggestions: newSuggestions,
+        clickHandler: newClickHandler
+      } = newProps;
+      if (
+        newIsVisible !== isVisible ||
+        newWord !== word ||
+        newScerenPos !== screenPos ||
+        newSuggestions !== suggestions ||
+        newClickHandler !== clickHandlerProp
+      ) {
+        setState(newProps);
+      }
+    });
     return unsubscribe;
-  });
+  }, [isVisible, word, screenPos, suggestions, clickHandlerProp]);
 
   const clickHandler = (event: React.MouseEvent) => {
     const target = event.target as HTMLElement;
