+++ src/lib/plugins/autocomplete/index.ts
@@ -1,5 +1,1 @@
-import Popup from "./popup";
-
 export { default } from "./autocomplete";
-
-export { Popup };
