--- src/lib/wrapper/index.ts
@@ -1,2 +1,0 @@
-export { default as useProseMirror } from "./useProseMirror";
-export { default as ProseMirror } from "./ProseMirror";
